[MET-34]search_assets endpoint (#1)
* search_assets endpoint implementation

before and after are used as proper previous_page and next_page tokens, and not as public keys of the assets. The correct way to paginate is to take the before/after from top of the response object and use it, and not a public key of the asset directly. Failing to do so and using the public key will not throw any errors, and will return some result, which might not be intended by the user.

* updated grouping handling as suggested

* additional refactoring