feat: overhaul task spawning in the whole project (#433)
* feat: overhaul task spawning in the whole project

Overhaul task spawning in all modules. Remove usage of
the `mutexed_tasks` concept, remove usage of a broadcast
channel that was used for shutdown, create a static
executor to save running tasks & exit gracefully.

* feat(nft_ingester): spawn graceful shutdown as the first task in `main`

Spawn graceful shutdown as the first tokio task in binaries that perform
it, await it at the end.

* feat(ingester,synchronizer): introduce more cancellation checks

* chore: remove patch versions from the ecosystem crates

* feat(ingester): replace some loops with `run_until_cancelled`